:recycle: refactor: Refatorada implementação de Pagamento

+++ src/main/java/br/dev/paulowolfgang/gestao_apolices/service/impl/PagamentoServiceImpl.java
@@ -1,12 +1,14 @@
 package br.dev.paulowolfgang.gestao_apolices.service.impl;
 
+import br.dev.paulowolfgang.gestao_apolices.dto.mapper.PagamentoMapper;
+import br.dev.paulowolfgang.gestao_apolices.dto.request.PagamentoRequestDto;
+import br.dev.paulowolfgang.gestao_apolices.dto.response.PagamentoResponseDto;
 import br.dev.paulowolfgang.gestao_apolices.entity.Pagamento;
 import br.dev.paulowolfgang.gestao_apolices.repository.IPagamentoRepository;
 import br.dev.paulowolfgang.gestao_apolices.service.IPagamentoService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class PagamentoServiceImpl implements IPagamentoService {
@@ -18,35 +20,40 @@ public class PagamentoServiceImpl implements IPagamentoService {
     }
 
     @Override
-    public Pagamento salvar(Pagamento pagamento) {
-        return pagamentoRepository.save(pagamento);
+    public PagamentoResponseDto salvar(PagamentoRequestDto request) {
+        Pagamento pagamento = PagamentoMapper.converter(request);
+        pagamento = pagamentoRepository.save(pagamento);
+        return PagamentoMapper.converter(pagamento);
     }
 
     @Override
-    public Optional<Pagamento> buscarPorId(Long id) {
-        return pagamentoRepository.findById(id);
+    public PagamentoResponseDto buscarPorId(Long id) {
+        Pagamento pagamento = pagamentoRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Pagamento não encontrado para o ID: " + id));
+        return PagamentoMapper.converter(pagamento);
     }
 
     @Override
-    public List<Pagamento> listarTodos() {
-        return pagamentoRepository.findAll();
+    public List<PagamentoResponseDto> listarTodos() {
+        List<Pagamento> pagamentos = pagamentoRepository.findAll();
+        return pagamentos.stream()
+                .map(PagamentoMapper::converter)
+                .toList();
     }
 
     @Override
-    public Pagamento atualizar(Long id, Pagamento pagamentoAtualizado) {
-        return pagamentoRepository.findById(id)
-                .map(pagamentoExistente -> {
-                    pagamentoExistente.setApolice(pagamentoAtualizado.getApolice());
-                    pagamentoExistente.setValor(pagamentoAtualizado.getValor());
-                    pagamentoExistente.setDataVencimento(pagamentoAtualizado.getDataVencimento());
-                    pagamentoExistente.setDataPagamento(pagamentoAtualizado.getDataPagamento());
-                    pagamentoExistente.setStatus(pagamentoAtualizado.getStatus());
-                    return pagamentoRepository.save(pagamentoExistente);
-                }).orElseThrow(() -> new IllegalArgumentException("Pagamento não encontrado para o ID: " + id));
+    public PagamentoResponseDto atualizar(Long id, PagamentoRequestDto pagamentoAtualizado) {
+        Pagamento pagamento = pagamentoRepository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Pagamento não encontrado para o ID: " + id));
+        PagamentoMapper.copiarParaPropriedades(pagamentoAtualizado, pagamento);
+        return PagamentoMapper.converter(pagamento);
     }
 
     @Override
     public void remover(Long id) {
+        if(!pagamentoRepository.existsById(id)){
+            throw new IllegalArgumentException("Pagamento não encontrado para o ID: " + id);
+        }
         pagamentoRepository.deleteById(id);
     }
 }
